pyproject.toml
[project]
name = "froide-pressconference"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "django<6",
    "django-cms",
    "lxml>=6.0.0",
    "pydoll-python>=2.20.1",
    "pyparsing[diagrams]>=3.3.2",
    "python-slugify>=8.0.4",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["froide_pressconference"]

[tool.ruff]
exclude = ["migrations", "build", "node_modules"]

[tool.ruff.lint]
ignore = ["E501", "C901"]
select = ["C", "E", "F", "W", "B", "I001"]

[tool.ruff.lint.pycodestyle]
max-line-length = 88

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]
default-section = "third-party"
known-first-party = ["froide"]

[tool.ruff.lint.isort.sections]
django = ["django"]

[tool.coverage.run]
source = ["froide_pressconference"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = ["pragma: no cover"]

[tool.djlint]
ignore = "T002,T003,H005,H006,H021,H023,H029,H030,H031"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "froide_pressconference.settings"
# addopts = ["--reuse-db"]
filterwarnings = [
    "ignore::DeprecationWarning:(?!froide_pressconference).*",
    "ignore::PendingDeprecationWarning:(?!froide_pressconference)*",
]

[dependency-groups]
dev = ["ipdb>=0.13.13", "pytest>=9.0.2"]
